requirements.txt
rich>=13.7.0
pyyaml>=6.0.1
docker>=7.0.0
click>=8.1.7
paramiko>=3.4.0
cryptography>=41.0.0
requests>=2.31.0